ai-signalement-review/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20

pillow==11.1.0
torch
detoxify
transformers